moved and renamed classes, paths in SeedData, refactored Order constructor to match seed data
