+++ orders/src/main/java/com/lambda/orders/models/Order.java
@@ -26,12 +26,12 @@ public class Order
   {
   }
 
-  public Order(double ordamount, double advanceamount, String orddescription, Customer customer)
+  public Order(double ordamount, double advanceamount, Customer customer, String orddescription)
   {
     this.ordamount = ordamount;
     this.advanceamount = advanceamount;
-    this.orddescription = orddescription;
     this.customer = customer;
+    this.orddescription = orddescription;
   }
 
   public long getOrdnum()
